changed name of MapInterface to MapManager; centralized usage of impl_system_param macro; updated collision.rs to use centralized macro;
